# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Biblioteca do Condomínio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List readers
      security:
        - adminBearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/UserStatus"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create reader
      security:
        - adminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get reader profile
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reader profile with loan history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update reader
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete reader
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Loans ───────────────────────────────────────────────────────────────────
...
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List reservations
      security:
        - adminBearerAuth: []
      parameters:
        - name: bookId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReservationStatus"
      responses:
        "200":
          description: Reservation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationsListResponse"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Add to reservation queue
      security:
        - adminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationDetail"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservations/{id}:
    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Execute reservation action
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationActionRequest"
      responses:
        "200":
          description: Action executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationDetail"
        "400":
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteReservation
      tags: [reservations]
      summary: Remove from queue
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /reservations/{id}/notify:
    patch:
      operationId: notifyReservation
      tags: [reservations]
      summary: Record WhatsApp notification sent
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationDetail"

  # ─── Reader Portal ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    readerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ─── Enums ─────────────────────────────────────────────────────────────────
    BookStatus:
      type: string
      enum: [draft, available, borrowed, reserved, lost, unavailable]
    UserStatus:
      type: string
      enum: [pending, active, inactive, blocked]
    LoanStatus:
      type: string
      enum: [active, returned, overdue]
    ReservationStatus:
      type: string
      enum: [waiting, notified, fulfilled, cancelled]
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        block:
          type: string
          nullable: true
        house:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/UserStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, status, createdAt, updatedAt]
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/UserProfile"
        - type: object
          properties:
            loans:
              type: array
              items:
                $ref: "#/components/schemas/LoanSummary"
          required: [loans]
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [users, meta]
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        block:
          type: string
          nullable: true
        house:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/UserStatus"
        password:
          type: string
          nullable: true
      required: [name, email]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        block:
          type: string
          nullable: true
        house:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/UserStatus"
        password:
          type: string
          nullable: true

    # ─── Books ─────────────────────────────────────────────────────────────────
    BookSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        isbn:
          type: string
          nullable: true
        publishedYear:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/BookStatus"
      required: [id, title, author, genre, status]
...
    LoanSummary:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        status:
          $ref: "#/components/schemas/LoanStatus"
        loanDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
          nullable: true
        book:
          $ref: "#/components/schemas/BookSummary"
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [id, bookId, userId, status, loanDate, dueDate]
...
    ReservationDetail:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        position:
          type: integer
        status:
          $ref: "#/components/schemas/ReservationStatus"
        notifiedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        book:
          $ref: "#/components/schemas/BookSummary"
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [id, bookId, userId, position, status, createdAt, updatedAt]
    ReservationsListResponse:
      type: object
      properties:
        reservations:
          type: array
          items:
            $ref: "#/components/schemas/ReservationDetail"
      required: [reservations]
    CreateReservationRequest:
      type: object
      properties:
        bookId:
          type: integer
        userId:
          type: integer
      required: [bookId, userId]
    ReservationActionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [notify, cancel, fulfill, advance]
      required: [action]

    # ─── Reader ────────────────────────────────────────────────────────────────
...