# lib/api-zod/src/generated/types/usersListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Biblioteca do Condomínio API
 * OpenAPI spec version: 1.0.0
 */
import type { PaginationMeta } from "./paginationMeta";
import type { UserProfile } from "./userProfile";

export interface UsersListResponse {
  users: UserProfile[];
  meta: PaginationMeta;
}